Cleanup, Refactor
- Refactor: Refactored the Keystone v2 service to use a base class for the logging functionality much like the models do
- Cleanup: stopped import logging where it's no longer necessary
