+++ openstackinabox/services/__init__.py
@@ -1,3 +1,4 @@
 """
 OpenStack Services
 """
+from openstackinabox.services.base_service import *
